Adjust file architecture & rename function & delete redundant comment
