--- python/tunnel_code.py
@@ -1,11 +1,0 @@
-from enum import Enum
-
-
-class TunnelCode(Enum):
-    PROBLEM = "problem"
-    TESTCASE = "testcase"
-    SOLUTION = "problem_solution"
-    CHECKER = "problem_checker"
-    USER_AVATER = "user_avater"
-    SUBMISSION = "user_submission"
-    USER_PROFILE = "user_profile"
